AR training: fp32 master weights + bf16 autocast; parallel latent preprocess
Precision: cfg.dtype is now the param/optimizer dtype (fp32 for the real
configs) and cfg.mixed_precision drives a derived autocast_dtype. The backbone
_call runs the transformer forward under torch.autocast(bf16) when set, keeping
fp32 master weights + fp32 AdamW state (stable at lr=6e-6 over 200k steps) with
bf16-speed compute; backward/optimizer step stay outside autocast. Score
functions cast to fp32 so the DMD score difference and CFG math run in fp32, and
the scheduler casts sigma to the latent dtype. autocast_dtype is None when params
already carry the compute dtype, so a dtype=bf16 smoke is unchanged.

Preprocess: scripts/preprocess_ar_latents.py gains --num-shards/--shard-id for an
embarrassingly-parallel SLURM-array re-encode and a CPU-only --merge pass that
concatenates per-shard sample lists and computes action-percentile stats.

Validated on H200 (validate_train_step.py): full self-forcing/DMD loop on real
Wan-1.3B, finite fp32 losses. Tests: 13 passed (adds autocast-dtype + scheduler
dtype-preservation regression tests).

+++ openworld/autoregressive/distill/dmd.py
@@ -32,10 +32,13 @@ def make_cfg_score_fn(
     """Wrap a backbone's ``forward_train`` as a (CFG'd) velocity score function."""
 
     def score_fn(x_sigma, timestep, cond):
-        v_cond = backbone.forward_train(x_sigma, timestep, cond, frames_per_block=frames_per_block)
+        # Cast scores to fp32: under bf16 autocast the backbone returns bf16, but
+        # the DMD score difference (x0_fake - x0_real) and the CFG combination
+        # difference two large, similar tensors — keep that arithmetic in fp32.
+        v_cond = backbone.forward_train(x_sigma, timestep, cond, frames_per_block=frames_per_block).float()
         if scale == 1.0 or null_cond is None:
             return v_cond
-        v_uncond = backbone.forward_train(x_sigma, timestep, null_cond, frames_per_block=frames_per_block)
+        v_uncond = backbone.forward_train(x_sigma, timestep, null_cond, frames_per_block=frames_per_block).float()
         return v_uncond + scale * (v_cond - v_uncond)
 
     return score_fn
